backend/requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.11.0
asyncpg==0.30.0
billiard==4.2.2
celery==5.5.3
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
fastapi==0.118.0
greenlet==3.2.4
h11==0.16.0
idna==3.10
kombu==5.5.4
loguru==0.7.3
packaging==25.0
prompt_toolkit==3.0.52
pydantic==2.11.10
pydantic-settings==2.11.0
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
redis==6.4.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.48.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uvicorn==0.37.0
vine==5.1.0
wcwidth==0.2.14
win32_setctime==1.2.0